requirements.txt
streamlit>=1.28.0
google-genai>=0.2.0
python-dotenv>=1.0.0
pyyaml>=6.0
